Move export and import functions out of main program
Just trying to clean up a little.
